Create fathersonage.java
Write a program that demonstrates the handling of exceptions in the inheritance tree. Create a base class called “Father” and a derived class called “Son” which extends the base class. In Father's class, implement a constructor that takes the age and throws the exception WrongAge( ) when the input age<0. In Son class, implement a constructor that cases both father and son’s age and throws an exception if the son’s age is >=father’s age.